Crate rusoto_cloudwatch

Amazon CloudWatch monitors your Amazon Web Services (AWS) resources and the applications you run on AWS in real time. You can use CloudWatch to collect and track metrics, which are the variables you want to measure for your resources and applications.

CloudWatch alarms send notifications or automatically change the resources you are monitoring based on rules that you define. For example, you can monitor the CPU usage and disk reads and writes of your Amazon EC2 instances. Then, use this data to determine whether you should launch additional instances to handle increased load. You can also use this data to stop under-used instances to save money.

In addition to monitoring the built-in metrics that come with AWS, you can monitor your own custom metrics. With CloudWatch, you gain system-wide visibility into resource utilization, application performance, and operational health.

If you're using the service, you're probably looking for CloudWatchClient and CloudWatch.

Structs

AlarmHistoryItem
Represents the history of a specific alarm.

CloudWatchClient
A client for the CloudWatch API.

DashboardEntry
Represents a specific dashboard.

DashboardValidationMessage
An error or warning for the operation.

Datapoint
Encapsulates the statistical data that CloudWatch computes from metric data.

DeleteAlarmsInput

DeleteDashboardsInput

DeleteDashboardsOutput

DescribeAlarmHistoryInput

DescribeAlarmHistoryOutput

DescribeAlarmsForMetricInput

DescribeAlarmsForMetricOutput

DescribeAlarmsInput

DescribeAlarmsOutput

Dimension
Expands the identity of a metric.

DimensionFilter
Represents filters for a dimension.

DisableAlarmActionsInput

EnableAlarmActionsInput

GetDashboardInput

GetDashboardOutput

GetMetricDataInput

GetMetricDataOutput

GetMetricStatisticsInput

GetMetricStatisticsOutput

GetMetricWidgetImageInput

GetMetricWidgetImageOutput

ListDashboardsInput

ListDashboardsOutput

ListMetricsInput

ListMetricsOutput

ListTagsForResourceInput

ListTagsForResourceOutput

MessageData
A message returned by the GetMetricDataAPI, including a code and a description.

Metric
Represents a specific metric.

MetricAlarm
Represents an alarm.

MetricDataQuery
This structure is used in both GetMetricData and PutMetricAlarm. The supported use of this structure is different for those two operations.

When used in GetMetricData, it indicates the metric data to return, and whether this call is just retrieving a batch set of data for one metric, or is performing a math expression on metric data. A single GetMetricData call can include up to 100 MetricDataQuery structures.

When used in PutMetricAlarm, it enables you to create an alarm based on a metric math expression. Each MetricDataQuery in the array specifies either a metric to retrieve, or a math expression to be performed on retrieved metrics. A single PutMetricAlarm call can include up to 20 MetricDataQuery structures in the array. The 20 structures can include as many as 10 structures that contain a MetricStat parameter to retrieve a metric, and as many as 10 structures that contain the Expression parameter to perform a math expression. Of those Expression structures, one must have True as the value for ReturnData. The result of this expression is the value the alarm watches.

Any expression used in a PutMetricAlarm operation must return a single time series. For more information, see Metric Math Syntax and Functions in the Amazon CloudWatch User Guide.

Some of the parameters of this structure also have different uses whether you are using this structure in a GetMetricData operation or a PutMetricAlarm operation. These differences are explained in the following parameter list.

MetricDataResult
A GetMetricData call returns an array of MetricDataResult structures. Each of these structures includes the data points for that metric, along with the timestamps of those data points and other identifying information.

MetricDatum
Encapsulates the information sent to either create a metric or add new values to be aggregated into an existing metric.

MetricStat
This structure defines the metric to be returned, along with the statistics, period, and units.

PutDashboardInput

PutDashboardOutput

PutMetricAlarmInput

PutMetricDataInput

SetAlarmStateInput

StatisticSet
Represents a set of statistics that describes a specific metric.

Tag
A key-value pair associated with a CloudWatch resource.

TagResourceInput

TagResourceOutput

UntagResourceInput

UntagResourceOutput

Enums

DeleteAlarmsError
Errors returned by DeleteAlarms

DeleteDashboardsError
Errors returned by DeleteDashboards

DescribeAlarmHistoryError
Errors returned by DescribeAlarmHistory

DescribeAlarmsError
Errors returned by DescribeAlarms

DescribeAlarmsForMetricError
Errors returned by DescribeAlarmsForMetric

DisableAlarmActionsError
Errors returned by DisableAlarmActions

EnableAlarmActionsError
Errors returned by EnableAlarmActions

GetDashboardError
Errors returned by GetDashboard

GetMetricDataError
Errors returned by GetMetricData

GetMetricStatisticsError
Errors returned by GetMetricStatistics

GetMetricWidgetImageError
Errors returned by GetMetricWidgetImage

ListDashboardsError
Errors returned by ListDashboards

ListMetricsError
Errors returned by ListMetrics

ListTagsForResourceError
Errors returned by ListTagsForResource

PutDashboardError
Errors returned by PutDashboard

PutMetricAlarmError
Errors returned by PutMetricAlarm

PutMetricDataError
Errors returned by PutMetricData

SetAlarmStateError
Errors returned by SetAlarmState

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

Traits

CloudWatch
Trait representing the capabilities of the CloudWatch API. CloudWatch clients implement this trait.