Struct rusoto_cloudwatch::CloudWatchClient

pub struct CloudWatchClient { /* fields omitted */ }

A client for the CloudWatch API.

Methods

impl CloudWatchClient

pub fn new(region: Region) -> CloudWatchClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CloudWatchClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,