Struct rusoto_cloudwatch::MetricAlarm

pub struct MetricAlarm {
    pub actions_enabled: Option<bool>,
    pub alarm_actions: Option<Vec<String>>,
    pub alarm_arn: Option<String>,
    pub alarm_configuration_updated_timestamp: Option<String>,
    pub alarm_description: Option<String>,
    pub alarm_name: Option<String>,
    pub comparison_operator: Option<String>,
    pub dimensions: Option<Vec<Dimension>>,
    pub evaluate_low_sample_count_percentile: Option<String>,
    pub evaluation_periods: Option<i64>,
    pub extended_statistic: Option<String>,
    pub insufficient_data_actions: Option<Vec<String>>,
    pub metric_name: Option<String>,
    pub namespace: Option<String>,
    pub ok_actions: Option<Vec<String>>,
    pub period: Option<i64>,
    pub state_reason: Option<String>,
    pub state_reason_data: Option<String>,
    pub state_updated_timestamp: Option<String>,
    pub state_value: Option<String>,
    pub statistic: Option<String>,
    pub threshold: Option<f64>,
    pub treat_missing_data: Option<String>,
    pub unit: Option<String>,
}

Represents an alarm.

Fields
actions_enabled: Option<bool>

Indicates whether actions should be executed during any changes to the alarm state.

alarm_actions: Option<Vec<String>>

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

alarm_arn: Option<String>

The Amazon Resource Name (ARN) of the alarm.

alarm_configuration_updated_timestamp: Option<String>

The time stamp of the last update to the alarm configuration.

alarm_description: Option<String>

The description of the alarm.

alarm_name: Option<String>

The name of the alarm.

comparison_operator: Option<String>

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

dimensions: Option<Vec<Dimension>>

The dimensions for the metric associated with the alarm.

evaluate_low_sample_count_percentile: Option<String> evaluation_periods: Option<i64>

The number of periods over which data is compared to the specified threshold.

extended_statistic: Option<String>

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

insufficient_data_actions: Option<Vec<String>>

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

metric_name: Option<String>

The name of the metric associated with the alarm.

namespace: Option<String>

The namespace of the metric associated with the alarm.

ok_actions: Option<Vec<String>>

The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

period: Option<i64>

The period, in seconds, over which the statistic is applied.

state_reason: Option<String>

An explanation for the alarm state, in text format.

state_reason_data: Option<String>

An explanation for the alarm state, in JSON format.

state_updated_timestamp: Option<String>

The time stamp of the last update to the alarm state.

state_value: Option<String>

The state value for the alarm.

statistic: Option<String>

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic.

threshold: Option<f64>

The value to compare with the specified statistic.

treat_missing_data: Option<String> unit: Option<String>

The unit of the metric associated with the alarm.