Struct rusoto_cloudwatch::MetricDatum

pub struct MetricDatum {
    pub dimensions: Option<Vec<Dimension>>,
    pub metric_name: String,
    pub statistic_values: Option<StatisticSet>,
    pub storage_resolution: Option<i64>,
    pub timestamp: Option<String>,
    pub unit: Option<String>,
    pub value: Option<f64>,
}

Encapsulates the information sent to either create a metric or add new values to be aggregated into an existing metric.

Fields
dimensions: Option<Vec<Dimension>>

The dimensions associated with the metric.

metric_name: String

The name of the metric.

statistic_values: Option<StatisticSet>

The statistical values for the metric.

storage_resolution: Option<i64>

Valid values are 1 and 60. Setting this to 1 specifies this metric as a high-resolution metric, so that CloudWatch stores the metric with sub-minute resolution down to one second. Setting this to 60 specifies this metric as a regular-resolution metric, which CloudWatch stores at 1-minute resolution. Currently, high resolution is available only for custom metrics. For more information about high-resolution metrics, see High-Resolution Metrics in the Amazon CloudWatch User Guide.

This field is optional, if you do not specify it the default of 60 is used.

timestamp: Option<String>

The time the metric data was received, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

unit: Option<String>

The unit of the metric.

value: Option<f64>

The value for the metric.

Although the parameter accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2). In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.