Struct rusoto_cloudwatch::PutMetricAlarmInput

pub struct PutMetricAlarmInput {
    pub actions_enabled: Option<bool>,
    pub alarm_actions: Option<Vec<String>>,
    pub alarm_description: Option<String>,
    pub alarm_name: String,
    pub comparison_operator: String,
    pub dimensions: Option<Vec<Dimension>>,
    pub evaluate_low_sample_count_percentile: Option<String>,
    pub evaluation_periods: i64,
    pub extended_statistic: Option<String>,
    pub insufficient_data_actions: Option<Vec<String>>,
    pub metric_name: String,
    pub namespace: String,
    pub ok_actions: Option<Vec<String>>,
    pub period: i64,
    pub statistic: Option<String>,
    pub threshold: f64,
    pub treat_missing_data: Option<String>,
    pub unit: Option<String>,
}

Fields
actions_enabled: Option<bool>

Indicates whether actions should be executed during any changes to the alarm state.

alarm_actions: Option<Vec<String>>

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

alarm_description: Option<String>

The description for the alarm.

alarm_name: String

The name for the alarm. This name must be unique within the AWS account.

comparison_operator: String

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

dimensions: Option<Vec<Dimension>>

The dimensions for the metric associated with the alarm.

evaluate_low_sample_count_percentile: Option<String>

Used only for alarms based on percentiles. If you specify ignore, the alarm state will not change during periods with too few data points to be statistically significant. If you specify evaluate or omit this parameter, the alarm will always be evaluated and possibly change state no matter how many data points are available. For more information, see Percentile-Based CloudWatch Alarms and Low Data Samples.

Valid Values: evaluate | ignore

evaluation_periods: i64

The number of periods over which data is compared to the specified threshold.

extended_statistic: Option<String>

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

insufficient_data_actions: Option<Vec<String>>

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

metric_name: String

The name for the metric associated with the alarm.

namespace: String

The namespace for the metric associated with the alarm.

ok_actions: Option<Vec<String>>

The actions to execute when this alarm transitions to an OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

period: i64

The period, in seconds, over which the specified statistic is applied.

statistic: Option<String>

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic.

threshold: f64

The value against which the specified statistic is compared.

treat_missing_data: Option<String>

Sets how this alarm is to handle missing data points. If TreatMissingData is omitted, the default behavior of missing is used. For more information, see Configuring How CloudWatch Alarms Treats Missing Data.

Valid Values: breaching | notBreaching | ignore | missing

unit: Option<String>

The unit of measure for the statistic. For example, the units for the Amazon EC2 NetworkIn metric are Bytes because NetworkIn tracks the number of bytes that an instance receives on all network interfaces. You can also specify a unit when you create a custom metric. Units help provide conceptual meaning to your data. Metric data points that specify a unit of measure, such as Percent, are aggregated separately.

If you specify a unit, you must use a unit that is appropriate for the metric. Otherwise, the Amazon CloudWatch alarm can get stuck in the INSUFFICIENT DATA state.