Trait rusoto_cloudwatch::CloudWatch

pub trait CloudWatch {
    fn delete_alarms(
        &self, 
        input: &DeleteAlarmsInput
    ) -> Result<(), DeleteAlarmsError>;
    fn describe_alarm_history(
        &self, 
        input: &DescribeAlarmHistoryInput
    ) -> Result<DescribeAlarmHistoryOutput, DescribeAlarmHistoryError>;
    fn describe_alarms(
        &self, 
        input: &DescribeAlarmsInput
    ) -> Result<DescribeAlarmsOutput, DescribeAlarmsError>;
    fn describe_alarms_for_metric(
        &self, 
        input: &DescribeAlarmsForMetricInput
    ) -> Result<DescribeAlarmsForMetricOutput, DescribeAlarmsForMetricError>;
    fn disable_alarm_actions(
        &self, 
        input: &DisableAlarmActionsInput
    ) -> Result<(), DisableAlarmActionsError>;
    fn enable_alarm_actions(
        &self, 
        input: &EnableAlarmActionsInput
    ) -> Result<(), EnableAlarmActionsError>;
    fn get_metric_statistics(
        &self, 
        input: &GetMetricStatisticsInput
    ) -> Result<GetMetricStatisticsOutput, GetMetricStatisticsError>;
    fn list_metrics(
        &self, 
        input: &ListMetricsInput
    ) -> Result<ListMetricsOutput, ListMetricsError>;
    fn put_metric_alarm(
        &self, 
        input: &PutMetricAlarmInput
    ) -> Result<(), PutMetricAlarmError>;
    fn put_metric_data(
        &self, 
        input: &PutMetricDataInput
    ) -> Result<(), PutMetricDataError>;
    fn set_alarm_state(
        &self, 
        input: &SetAlarmStateInput
    ) -> Result<(), SetAlarmStateError>;
}

Trait representing the capabilities of the CloudWatch API. CloudWatch clients implement this trait.

Required Methods

fn delete_alarms(
&self,
input: &DeleteAlarmsInput
) -> Result<(), DeleteAlarmsError>

Deletes the specified alarms. In the event of an error, no alarms are deleted.

fn describe_alarm_history(
&self,
input: &DescribeAlarmHistoryInput
) -> Result<DescribeAlarmHistoryOutput, DescribeAlarmHistoryError>

Retrieves the history for the specified alarm. You can filter the results by date range or item type. If an alarm name is not specified, the histories for all alarms are returned.

Note that Amazon CloudWatch retains the history of an alarm even if you delete the alarm.

fn describe_alarms(
&self,
input: &DescribeAlarmsInput
) -> Result<DescribeAlarmsOutput, DescribeAlarmsError>

Retrieves the specified alarms. If no alarms are specified, all alarms are returned. Alarms can be retrieved by using only a prefix for the alarm name, the alarm state, or a prefix for any action.

fn describe_alarms_for_metric(
&self,
input: &DescribeAlarmsForMetricInput
) -> Result<DescribeAlarmsForMetricOutput, DescribeAlarmsForMetricError>

Retrieves the alarms for the specified metric. Specify a statistic, period, or unit to filter the results.

fn disable_alarm_actions(
&self,
input: &DisableAlarmActionsInput
) -> Result<(), DisableAlarmActionsError>

Disables the actions for the specified alarms. When an alarm's actions are disabled, the alarm actions do not execute when the alarm state changes.

fn enable_alarm_actions(
&self,
input: &EnableAlarmActionsInput
) -> Result<(), EnableAlarmActionsError>

Enables the actions for the specified alarms.

fn get_metric_statistics(
&self,
input: &GetMetricStatisticsInput
) -> Result<GetMetricStatisticsOutput, GetMetricStatisticsError>

Gets statistics for the specified metric.

Amazon CloudWatch retains metric data as follows:

Data points with a period of 60 seconds (1 minute) are available for 15 days

Data points with a period of 300 seconds (5 minute) are available for 63 days

Data points with a period of 3600 seconds (1 hour) are available for 455 days (15 months)

Note that CloudWatch started retaining 5-minute and 1-hour metric data as of 9 July 2016.

The maximum number of data points returned from a single call is 1,440. If you request more than 1,440 data points, Amazon CloudWatch returns an error. To reduce the number of data points, you can narrow the specified time range and make multiple requests across adjacent time ranges, or you can increase the specified period. A period can be as short as one minute (60 seconds). Note that data points are not returned in chronological order.

Amazon CloudWatch aggregates data points based on the length of the period that you specify. For example, if you request statistics with a one-hour period, Amazon CloudWatch aggregates all data points with time stamps that fall within each one-hour period. Therefore, the number of values aggregated by CloudWatch is larger than the number of data points returned.

CloudWatch needs raw data points to calculate percentile statistics. If you publish data using a statistic set instead, you cannot retrieve percentile statistics for this data unless one of the following conditions is true:

The SampleCount of the statistic set is 1

The Min and the Max of the statistic set are equal

For a list of metrics and dimensions supported by AWS services, see the Amazon CloudWatch Metrics and Dimensions Reference in the Amazon CloudWatch User Guide.

fn list_metrics(
&self,
input: &ListMetricsInput
) -> Result<ListMetricsOutput, ListMetricsError>

List the specified metrics. You can use the returned metrics with GetMetricStatistics to obtain statistical data.

Up to 500 results are returned for any one call. To retrieve additional results, use the returned token with subsequent calls.

After you create a metric, allow up to fifteen minutes before the metric appears. Statistics about the metric, however, are available sooner using GetMetricStatistics.

fn put_metric_alarm(
&self,
input: &PutMetricAlarmInput
) -> Result<(), PutMetricAlarmError>

Creates or updates an alarm and associates it with the specified metric. Optionally, this operation can associate one or more Amazon SNS resources with the alarm.

When this operation creates an alarm, the alarm state is immediately set to INSUFFICIENT_DATA. The alarm is evaluated and its state is set appropriately. Any actions associated with the state are then executed.

When you update an existing alarm, its state is left unchanged, but the update completely overwrites the previous configuration of the alarm.

If you are an AWS Identity and Access Management (IAM) user, you must have Amazon EC2 permissions for some operations:

ec2:DescribeInstanceStatus and ec2:DescribeInstances for all alarms on EC2 instance status metrics

ec2:StopInstances for alarms with stop actions

ec2:TerminateInstances for alarms with terminate actions

ec2:DescribeInstanceRecoveryAttribute and ec2:RecoverInstances for alarms with recover actions

If you have read/write permissions for Amazon CloudWatch but not for Amazon EC2, you can still create an alarm, but the stop or terminate actions won't be performed. However, if you are later granted the required permissions, the alarm actions that you created earlier will be performed.

If you are using an IAM role (for example, an Amazon EC2 instance profile), you cannot stop or terminate the instance using alarm actions. However, you can still see the alarm state and perform any other actions such as Amazon SNS notifications or Auto Scaling policies.

If you are using temporary security credentials granted using the AWS Security Token Service (AWS STS), you cannot stop or terminate an Amazon EC2 instance using alarm actions.

Note that you must create at least one stop, terminate, or reboot alarm using the Amazon EC2 or CloudWatch console to create the EC2ActionsAccess IAM role. After this IAM role is created, you can create stop, terminate, or reboot alarms using a command-line interface or an API.

fn put_metric_data(
&self,
input: &PutMetricDataInput
) -> Result<(), PutMetricDataError>

Publishes metric data points to Amazon CloudWatch. Amazon CloudWatch associates the data points with the specified metric. If the specified metric does not exist, Amazon CloudWatch creates the metric. When Amazon CloudWatch creates a metric, it can take up to fifteen minutes for the metric to appear in calls to ListMetrics.

Each PutMetricData request is limited to 40 KB in size for HTTP POST requests.

Although the Value parameter accepts numbers of type Double, Amazon CloudWatch rejects values that are either too small or too large. Values must be in the range of 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2). In addition, special values (e.g., NaN, +Infinity, -Infinity) are not supported.

You can use up to 10 dimensions per metric to further clarify what data the metric collects. For more information on specifying dimensions, see Publishing Metrics in the Amazon CloudWatch User Guide.

Data points with time stamps from 24 hours ago or longer can take at least 48 hours to become available for GetMetricStatistics from the time they are submitted.

CloudWatch needs raw data points to calculate percentile statistics. If you publish data using a statistic set instead, you cannot retrieve percentile statistics for this data unless one of the following conditions is true:

The SampleCount of the statistic set is 1

The Min and the Max of the statistic set are equal

fn set_alarm_state(
&self,
input: &SetAlarmStateInput
) -> Result<(), SetAlarmStateError>

Temporarily sets the state of an alarm for testing purposes. When the updated state differs from the previous value, the action configured for the appropriate state is invoked. For example, if your alarm is configured to send an Amazon SNS message when an alarm is triggered, temporarily changing the alarm state to ALARM sends an Amazon SNS message. The alarm returns to its actual state (often within seconds). Because the alarm state change happens very quickly, it is typically only visible in the alarm's History tab in the Amazon CloudWatch console or through DescribeAlarmHistory.