Crate rusoto_cloudwatch

Amazon CloudWatch monitors your Amazon Web Services (AWS) resources and the applications you run on AWS in real-time. You can use CloudWatch to collect and track metrics, which are the variables you want to measure for your resources and applications.

CloudWatch alarms send notifications or automatically make changes to the resources you are monitoring based on rules that you define. For example, you can monitor the CPU usage and disk reads and writes of your Amazon Elastic Compute Cloud (Amazon EC2) instances and then use this data to determine whether you should launch additional instances to handle increased load. You can also use this data to stop under-used instances to save money.

In addition to monitoring the built-in metrics that come with AWS, you can monitor your own custom metrics. With CloudWatch, you gain system-wide visibility into resource utilization, application performance, and operational health.

If you're using the service, you're probably looking for CloudWatchClient and CloudWatch.

Structs

AlarmHistoryItem
Represents the history of a specific alarm.

CloudWatchClient
A client for the CloudWatch API.

Datapoint
Encapsulates the statistical data that Amazon CloudWatch computes from metric data.

DeleteAlarmsInput

DescribeAlarmHistoryInput

DescribeAlarmHistoryOutput

DescribeAlarmsForMetricInput

DescribeAlarmsForMetricOutput

DescribeAlarmsInput

DescribeAlarmsOutput

Dimension
Expands the identity of a metric.

DimensionFilter
Represents filters for a dimension.

DisableAlarmActionsInput

EnableAlarmActionsInput

GetMetricStatisticsInput

GetMetricStatisticsOutput

ListMetricsInput

ListMetricsOutput

Metric
Represents a specific metric.

MetricAlarm
Represents an alarm.

MetricDatum
Encapsulates the information sent to either create a metric or add new values to be aggregated into an existing metric.

PutMetricAlarmInput

PutMetricDataInput

SetAlarmStateInput

StatisticSet
Represents a set of statistics that describes a specific metric.

Enums

DeleteAlarmsError
Errors returned by DeleteAlarms

DescribeAlarmHistoryError
Errors returned by DescribeAlarmHistory

DescribeAlarmsError
Errors returned by DescribeAlarms

DescribeAlarmsForMetricError
Errors returned by DescribeAlarmsForMetric

DisableAlarmActionsError
Errors returned by DisableAlarmActions

EnableAlarmActionsError
Errors returned by EnableAlarmActions

GetMetricStatisticsError
Errors returned by GetMetricStatistics

ListMetricsError
Errors returned by ListMetrics

PutMetricAlarmError
Errors returned by PutMetricAlarm

PutMetricDataError
Errors returned by PutMetricData

SetAlarmStateError
Errors returned by SetAlarmState

Traits

CloudWatch
Trait representing the capabilities of the CloudWatch API. CloudWatch clients implement this trait.